move globals/macros to inst/include/RcppSimdJson/common.hpp and add documentation
